Create python test action
On Github using creating a chosen "Action" creates a "Workflow" (python-app.yml).
It's configured to run pylint and pytest.

Essential what is happening is on whatever trigger that is happening to the repository ("on:" tag, -> here the triggers are on any "push " action and any "pull_request" )
Github goes through .github/workflows/python-app.yml and runs a compose on it. On some GitHub server I suppose, it spins up a container that does some operations described in the .yml.

In this yml it is described to run a python container, install dependancies and run pytest and pylint.